backend/requirements.txt
fastapi[standard]
uvicorn
requests
boto3
python-multipart
pillow
openai